+++ src/main/kotlin/com/rickbusarow/ktrules/rules/KDocLeadingAsteriskRule.kt
@@ -26,7 +26,6 @@ import com.rickbusarow.ktrules.rules.internal.psi.isKDocWhitespaceBeforeLeadingA
 import com.rickbusarow.ktrules.rules.internal.psi.isWhiteSpace
 import com.rickbusarow.ktrules.rules.internal.psi.isWhiteSpaceWithNewline
 import com.rickbusarow.ktrules.rules.internal.psi.nextLeaf
-import com.rickbusarow.ktrules.rules.internal.psi.nextSibling
 import com.rickbusarow.ktrules.rules.internal.psi.parent
 import org.jetbrains.kotlin.com.intellij.lang.ASTNode
 import org.jetbrains.kotlin.com.intellij.psi.impl.source.tree.LeafPsiElement
@@ -38,14 +37,14 @@ import org.jetbrains.kotlin.kdoc.psi.api.KDoc
  *
  * @since 1.0.1
  */
-class KDocLeadingAsteriskRule : RuleCompat(ID,) {
+class KDocLeadingAsteriskRule : RuleCompat(ID) {
   override fun beforeVisitChildNodes(
     node: ASTNode,
     autoCorrect: Boolean,
     emit: (offset: Int, errorMessage: String, canBeAutoCorrected: Boolean) -> Unit
   ) {
 
-    if (node.elementType == ElementType.KDOC_START) {
+    if (node.elementType == ElementType.KDOC) {
       visitKDoc(node, autoCorrect = autoCorrect, emit = emit)
     }
   }
@@ -56,56 +55,69 @@ class KDocLeadingAsteriskRule : RuleCompat(ID,) {
     emit: (offset: Int, errorMessage: String, canBeAutoCorrected: Boolean) -> Unit
   ) {
 
-    val kdoc = kdocNode.psi.parent as KDoc
+    val kdoc = kdocNode.psi as KDoc
 
     val indent = kdoc.fileIndent(additionalOffset = 1)
     val newlineIndent = "\n$indent"
 
-    kdoc.node.childrenDepthFirst()
-      .distinct()
+    kdocNode.childrenDepthFirst()
       .filter { it.elementType == ElementType.WHITE_SPACE }
       .filter { it.text.lines().size > 1 }
-      .filter { !it.isKDocWhitespaceBeforeLeadingAsterisk() }
+      .filter { !(it.isKDocWhitespaceBeforeLeadingAsterisk() && it.text.lines().size == 2) }
       .filter { node ->
         val next = node.nextLeaf()
+        val nodeLines = node.text.lines()
 
         when {
           next == null -> false
           next.isWhiteSpace() -> false
-          next.isKDocLeadingAsterisk() -> false
-          next.isKDocEnd() && node.text.lines().count() <= 2 -> false
+          next.isKDocLeadingAsterisk() && nodeLines.size <= 2 -> false
+          next.isKDocEnd() && nodeLines.size <= 2 -> false
           else -> true
         }
       }
       .toList()
       .forEach { node ->
 
-        val parent = node.parent!!
-
-        val next = node.nextSibling() ?: return@forEach
+        val next = node.nextLeaf(true) ?: return@forEach
 
         emit(next.startOffset, ERROR_MESSAGE, true)
 
         if (autoCorrect) {
 
-          val numLines = node.text.split("\n").size - 1
-
-          parent.removeChild(node)
-
-          repeat(numLines) { i ->
-            if (i == numLines - 1 && next.elementType == ElementType.KDOC_END) {
-              parent.addChild(PsiWhiteSpaceImpl(newlineIndent.dropLast(1)), next)
-            } else {
-              parent.addChild(PsiWhiteSpaceImpl(newlineIndent), next)
-              parent.addChild(LeafPsiElement(ElementType.KDOC_LEADING_ASTERISK, "*"), next)
-            }
-          }
-
-          if (!next.isWhiteSpaceWithNewline()) {
-            parent.addChild(PsiWhiteSpaceImpl(" "), next)
-          }
+          node.fixBlankLines(next, newlineIndent)
         }
       }
+  }
+
+  private fun ASTNode.fixBlankLines(next: ASTNode, newlineIndent: String) {
+    val numLines = text.count { it == '\n' }
+
+    val nodeParent = parent!!
+    val nextParent = next.parent!!
+
+    nodeParent.removeChild(this)
+
+    (1..numLines).forEach { i ->
+      when {
+        i == numLines && next.isKDocEnd() -> {
+          nextParent.addChild(PsiWhiteSpaceImpl(newlineIndent), next)
+        }
+
+        i == numLines && next.isKDocLeadingAsterisk() -> {
+          nextParent.addChild(PsiWhiteSpaceImpl(newlineIndent.dropLast(1)), next)
+        }
+
+        else -> {
+          nextParent.addChild(PsiWhiteSpaceImpl(newlineIndent), next)
+          nextParent.addChild(LeafPsiElement(ElementType.KDOC_LEADING_ASTERISK, "*"), next)
+        }
+      }
+    }
+
+    if (!next.isWhiteSpaceWithNewline() && !next.isKDocEnd()) {
+      nextParent.addChild(PsiWhiteSpaceImpl(" "), next)
+    }
   }
 
   internal companion object {